refactor(tl): add new translation.rs
